Added a div with id=schemafw-meta to the output generated by sfwtemplates.xsl.  The id makes it easy to find, because the data can't go in the <head> as a meta element, or in the <html> as a data- attribute because these are output by default.xsl, which must be dead simple.
The schemafw-meta element is only generated when a document
is processed as a full page, and it gives the framework a
chance to perceive, among other things, if the response is
from a post request, in which case the framework WILL NOT
try to get a cached copy of the XML document.  This is to
prevent the framework, in the event of a form request that
results in a new page view, from reposting a form request as
an empty GET request that obliterates the initial POSTed form.

Changes to sfwmain.js detect the schemafw-meta element and
process accordingly.